Add files via upload
Here are the main LLM data, as well as the fine-tuning and transfer learning data, in the paper "Teaching AI to Handle Exceptions: Supervised Fine-Tuning with Human-Aligned Judgment". Data is split and compressed into three files to allow for upload.

There are 9 columns:
title: official scenario name (i.e., Grocery prices).
scenario: scenario code name (i.e., grocery_prices).
level: 1 through 6, or l1 through l6.
system_message: initial context-setting message for some models; not all rows will have this.
prompt: prompt a model receives; not all rows will have this.
response: full model response.
model: name of the model being prompted. If it starts with binary_ or explanations_, it's fine-tuned on yes/no or full explanations, respectively. 
fine-tuned: 0 if the model is off the shelf and 1 if it's fine-tuned
transfer: 1 if this is transfer learning, 0 if it's a fine-tuned model applied to the same scenario and NA if it's not a fine-tuned model. This can also be deter,omed from the model and scenario values